Refactor enums to use enum class where possible, and to have an explicit base integral type. Flags enums are still in namespaces, but have a typedef value_type for the field, and standard enum ids Flag and FlagBit. Array index enums are still in namespaces.

+++ MWSE/xGetSpellInfo.cpp
@@ -61,7 +61,7 @@ namespace mwse
 		TES3::Spell* spell = tes3::getObjectById<TES3::Spell>(spellId, TES3::ObjectType::Spell);;
 		if (spell != NULL) {
 			name = spell->name;
-			type = spell->castType;
+			type = long(spell->castType);
 			cost = spell->magickaCost;
 			effects = tes3::getEffectCount(spell->effects);
 			flags = spell->spellFlags;
